# sdk/log: Add WithAllowKeyDuplication logger provider option