--- a/sdk/log/logger.go
+++ b/sdk/log/logger.go
@@ -94,6 +94,7 @@ func (l *logger) newRecord(ctx context.Context, r log.Record) Record {
 		scope:                     &l.instrumentationScope,
 		attributeValueLengthLimit: l.provider.attributeValueLengthLimit,
 		attributeCountLimit:       l.provider.attributeCountLimit,
+		allowDupAttrs:             l.provider.attributeAllowDuplicates,
 	}
 
 	// This field SHOULD be set once the event is observed by OpenTelemetry.