--- a/sdk/log/logger.go
+++ b/sdk/log/logger.go
@@ -94,7 +94,7 @@ func (l *logger) newRecord(ctx context.Context, r log.Record) Record {
 		scope:                     &l.instrumentationScope,
 		attributeValueLengthLimit: l.provider.attributeValueLengthLimit,
 		attributeCountLimit:       l.provider.attributeCountLimit,
-		allowDupKeys:              l.provider.keysAllowDuplicates,
+		allowDupKeys:              l.provider.allowDupKeys,
 	}
 
 	// This field SHOULD be set once the event is observed by OpenTelemetry.