--- a/sdk/log/provider.go
+++ b/sdk/log/provider.go
@@ -258,7 +258,7 @@ func WithAttributeValueLengthLimit(limit int) LoggerProviderOption {
 	})
 }
 
-// AllowKeyDuplication sets whether deduplication is skipped for log attributes or other key-value collections.
+// WithAllowKeyDuplication sets whether deduplication is skipped for log attributes or other key-value collections.
 //
 // By default, the key-value collections within a log record are deduplicated to comply with the OpenTelemetry Specification.
 // Deduplication means that if multiple key–value pairs with the same key are present, only a single pair
@@ -269,7 +269,7 @@ func WithAttributeValueLengthLimit(limit int) LoggerProviderOption {
 // Note that if you disable deduplication, you are responsible for ensuring that duplicate
 // key-value pairs within in a single collection are not emitted,
 // or that the telemetry receiver can handle such duplicates.
-func AllowKeyDuplication() LoggerProviderOption {
+func WithAllowKeyDuplication() LoggerProviderOption {
 	return loggerProviderOptionFunc(func(cfg providerConfig) providerConfig {
 		cfg.allowDupKeys = newSetting(true)
 		return cfg