--- a/sdk/log/provider_test.go
+++ b/sdk/log/provider_test.go
@@ -115,7 +115,7 @@ func TestNewLoggerProviderConfiguration(t *testing.T) {
 				WithProcessor(p1),
 				WithAttributeCountLimit(attrCntLim),
 				WithAttributeValueLengthLimit(attrValLenLim),
-				AllowKeyDuplication(),
+				WithAllowKeyDuplication(),
 			},
 			want: &LoggerProvider{
 				resource:                  res,